fix: refresh lens session
